# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borland Groover Appointment Setter Reference API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SearchResult {
  section: string;
  title: string;
  content: string;
  tags: string[];
}

export interface Note {
  id: number;
  content: string;
  createdAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  content: string;
}

export type SearchContentParams = {
/**
 * Search query
 * @minLength 1
 */
q: string;
};