# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Borland Groover Appointment Setter Reference API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: search
    description: Global content search
  - name: notes
    description: Personal agent notes
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /search:
    get:
      operationId: searchContent
      tags: [search]
      summary: Search all cheat sheet content
      description: Full-text search across all reference content sections
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Search query
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"

  /notes:
    get:
      operationId: getNotes
      tags: [notes]
      summary: Get all personal notes
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a personal note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/{id}:
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a personal note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SearchResult:
      type: object
      properties:
        section:
          type: string
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - section
        - title
        - content
        - tags

    Note:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - content
        - createdAt

    NoteInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required:
        - content