# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borland Groover Appointment Setter Reference API
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: number;
  content: string;
  createdAt: string;
}