# lib/api-zod/src/generated/types/noteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borland Groover Appointment Setter Reference API
 * OpenAPI spec version: 0.1.0
 */

export interface NoteInput {
  /** @minLength 1 */
  content: string;
}