# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borland Groover Appointment Setter Reference API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  section: string;
  title: string;
  content: string;
  tags: string[];
}